Update startup script, dockerfile
Use exec form for entrypoint, pass cmd arguments on to ozwdaemon, use
exec to launch ozwdaemon to drop the shell to allow signals to be passed
on to the application.
Update startup script to allow whitespace in parameters, only allow
"true" and "false" as valid options for STOP_ON_FAILURE and MQTT_TLS.
Re-add config-dir, user-dir configuration.