Feature: Improved startup behavior when MQTT broker is unavailable (#120) (#121)
* Delay ozwdaemon startup until MQTT broker is available

Wait until the MQTT broker port is available before starting up
ozwdaemon. This reduces the churning of the daemon starting then dying
while MQTT is offline.

The script will exit with error 124 if the MQTT broker port is
unavailable after a configurable amount of time (default 30 seconds).

* Use same startup script for standalone and allinone docker images

The scripts were identical so only one is necessary. Reduces
double-edits.

* Copy startozwdaemon.sh to /opt

The allinone image copies this script to /opt, so do the same with the
standalone image for consistency.

* Remove duplicate env vars

The entrypoint script defines the necessary environment variables, so
there's no need to duplicate them in the Dockerfile.

* Add documentation for MQTT_CONNECT_TIMEOUT env var

Improve docs for other environment variables.