Prevent docker build layer cache invalidation (#133)
The Dockerfile is including all files in the project, which means
touching *any* file in the project (README, docs, docker root files,
etc) invalidates practically all of the cached docker layers and
triggers a rebuild of all libraries and applications.

Manually copy the needed files instead of the entire project. Note that
using .dockerignore alone is not possible because it would prevent
copying some files into the image that are necessary but unrelated to
the application builds.